New styling + better founder creation handling

+++ app/Http/Controllers/Auth/RegisteredUserController.php
@@ -26,7 +26,8 @@ class RegisteredUserController extends Controller
      */
     public function create(): View
     {
-        return view('auth.register');
+        $firstuser = User::count() === 0;
+        return view('auth.register', ["firstuser" => $firstuser]);
     }
 
     /**
@@ -36,18 +37,21 @@ class RegisteredUserController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
-        $request->validate([
+        $firstuser = User::count() === 0;
+        $params = [
             'name' => ['required', 'string', 'max:255', 'unique:'.User::class, 'regex:/^[\w-]*$/'],
             'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
             'password' => ['required', 'confirmed', Rules\Password::defaults()],
-			'invite_code' => ['required', 'string', 'size:10'],
-        ]);
+        ];
+        if(!$firstuser) $params['invite_code'] = ['required', 'string', 'size:10'];
+        
+        $request->validate($params);
 		
 		DB::beginTransaction();
 		try {
-			$invite = Invite::where("id", trim($request->invite_code))->firstOrFail();
-			$inviterID = $invite->creator_id;
-			$invite->delete();
+			$invite = Invite::where("id", trim($request->invite_code))->first();
+			$inviterID = $invite ? $invite->creator_id : null;
+			if($invite) $invite->delete();
 
 			$user = User::create([
 				'name' => $request->name,
@@ -60,7 +64,6 @@ class RegisteredUserController extends Controller
 		} catch (Throwable $err) {
 			return redirect()->back()->withErrors("Account creation failed; try again with the same invite code. ".$err->getMessage());
 		}
-
 		DB::commit();
 
         event(new Registered($user));
